First attempt at automated build on Windows
The Windows build makes use of the same configuration script as the
Linux build, this is accomplished by using the copy of bash installed
with the Windows version of Git. The build is done using MinGW project
gcc and g++ for the compilers and Ninja as the build tool.

The requirements to build the project on Windows are:
 - Git for Windows (w/ Bash, added to the path)
 - MinGW x86_64 (in the path)
 - Ninja (also in the path)